separate sqlachemy converter testing
