+++ pyproject.toml
@@ -12,12 +12,12 @@ python = ">=3.8,<4"
 jsonschema = ">=3.0.0"
 parameterized = "0.7.*"
 PyYAML = ">=5.4.1"
-dataengineeringutils3 = ">=1.4.0"
 pyarrow = { version = ">=3.0.0", optional = true }
 etl-manager = { version = "^7.4", optional = true }
 psycopg2 = { version = "^2.9.2", optional = true,  extras = ["postgres"]}
 "testing.postgresql" ={ version = "^1.3.0", optional = true,  extras = ["postgres"]}
 SQLAlchemy = { version = "^1.4.27", optional = true,  extras = ["postgres"]}
+dataengineeringutils3 = ">=1.4.0"
 
 [tool.poetry.dev-dependencies]
 flake8 = "^3.8.4"
@@ -27,12 +27,12 @@ pytest-watch = ">=4.2.0"
 moto = "^2.2.11"
 pandas = "^1.5.0"
 duckdb-engine = "0.7.0"
+oracledb = "^1.3.0"
 
 [tool.poetry.extras]
 arrow = ["pyarrow"]
 etl-manager = ["etl-manager"]
 postgres = ["psycopg2","testing.postgresql","SQLAlchemy"]
-sqlalchemy = ["SQLAlchemy"]
 
 [build-system]
 requires = ["poetry>=0.12"]
