--- tests/.oracle_credentials.sh
@@ -1,3 +1,0 @@
-export DB_USER=my_user
-export DB_PASSWORD=my_password
-export PORT=1521
