+++ tests/scripts/.oracle_image.sh
@@ -1,5 +1,5 @@
 docker run -d --name oracle_db \
--p 1521:$PORT \
+-p 1521:1521 \
 -e ORACLE_RANDOM_PASSWORD="y" \
 -e APP_USER=$DB_USER \
 -e APP_USER_PASSWORD=$DB_PASSWORD \
