+++ tests/scripts/.sqlalchemy_envs.sh
@@ -1,0 +1,5 @@
+export DB_USER=my_user
+export DB_PASSWORD=my_password
+export TEST_ORACLE=True
+export TEST_SQLITE=True
+export TEST_POSTGRES=True
