+++ tests/test_sqlalchemy_converter.py
@@ -2,28 +2,15 @@ import sys
 import os
 import logging
 import pytest
-import oracledb
-from sqlalchemy import Column, Table, MetaData, create_engine, inspect, text
+from sqlalchemy import Column, Table, MetaData, create_engine
 from sqlalchemy.types import (
     String,
-    NCHAR,
-    VARCHAR,
-    CLOB,
     Integer,
-    BIGINT,
     Float,
     DECIMAL,
     NUMERIC,
-    DateTime,
-    Date,
-    TIMESTAMP,
-    DATE,
-    DATETIME,
     Boolean,
     LargeBinary,
-    BLOB,
-    VARBINARY,
-    JSON,
 )
 from mojap_metadata.converters.sqlalchemy_converter import SQLAlchemyConverter
 
@@ -35,28 +22,38 @@ from mojap_metadata.converters.sqlalchemy_converter import SQLAlchemyConverter
 
 logging.basicConfig(filename="db.log")
 logging.getLogger("sqlalchemy.engine").setLevel(logging.ERROR)
-
-# create oracle engine
-oracledb.version = "8.3.0"
-sys.modules["cx_Oracle"] = oracledb
+table_name = "my_table"
 user = os.getenv("DB_USER")
 password = os.getenv("DB_PASSWORD")
-port = os.getenv("PORT")
-oracle_engine = create_engine(
-    f"oracle://{user}:{password}@localhost:{port}/?service_name=XEPDB1"
-)
 
-sqlite_engine = create_engine("sqlite:///:memory:")
-duckdb_engine = create_engine("duckdb:///:memory:")
-table_name = "my_table"
-schema_name = user  # schema name needs to be the same as user name for Oracle
+# The following dialects are tested by default, skipped if value is different to 'True'
+test_sqlite = os.getenv("TEST_SQLITE", "True") == "True"
+test_duckdb = os.getenv("TEST_DUCKDB", "True") == "True"
+
+# The following dialects are skipped by default, tested if value is set to 'True'
+test_oracle = os.getenv("TEST_ORACLE", "False") == "True"
+test_postgres = os.getenv("TEST_POSTGRES", "False") == "True"
 
 
-def create_schema(connectable, schema_command, schema):
-    inspector = inspect(connectable)
-    if schema.lower() not in inspector.get_schema_names():
-        with connectable.connect() as conn:
-            conn.execute(text(f'CREATE {schema_command} "{schema}"'))
+# Create engines
+def create_sqlachemy_engine(dialect):
+    engine = None
+    if dialect == "sqlite":
+        engine = create_engine("sqlite:///:memory:")
+    elif dialect == "duckdb":
+        engine = create_engine("duckdb:///:memory:")
+    elif dialect == "oracle":
+        import oracledb
+        oracledb.version = "8.3.0"
+        sys.modules["cx_Oracle"] = oracledb
+        engine = create_engine(
+            f"oracle://{user}:{password}@localhost:1521/?service_name=XEPDB1"
+        )
+    elif dialect == "postgres":
+        engine = create_engine(
+            f"postgresql://postgres:{password}@localhost:5432/postgres"
+        )
+    return engine
 
 
 def create_tables(connectable, schema):
@@ -157,74 +154,65 @@ mojap_metadata/v1.3.0.json",
 
 
 @pytest.mark.parametrize(
-    "connectable,schema_command,schema,table_description,column_description,"
+    "dialect,schema,table_description,column_description,"
     "primary_key,float_type,bool_type",
     [
-        (
-            sqlite_engine,
-            "SCHEMA",
+        pytest.param(
+            "sqlite",
             "main",
             "",
             "",
             ["my_string_10"],
             "float64",
             "bool",
+            marks=(pytest.mark.skipif(test_sqlite is False, reason="skip sqlite test")),
         ),
-        (
-            duckdb_engine,
-            "SCHEMA",
+        pytest.param(
+            "duckdb",
             "main",
             "",
             "",
             [],
             "float16",
             "bool",
+            marks=(pytest.mark.skipif(test_duckdb is False, reason="skip duckdb test")),
         ),
-        (
-            "postgres_engine",
-            "SCHEMA",
-            schema_name,
+        pytest.param(
+            "postgres",
+            "public",
             "this is my table",
             "this is the comment",
             ["my_string_10"],
             "float16",
             "bool",
+            marks=(
+                pytest.mark.skipif(test_postgres is False, reason="skip postgres test")
+            ),
         ),
-        (
-            oracle_engine,
-            "USER",
-            schema_name,
+        pytest.param(
+            "oracle",
+            user,
             "this is my table",
             "this is the comment",
             ["my_string_10"],
             "float64",
             "int32",
+            marks=(pytest.mark.skipif(test_oracle is False, reason="skip oracle test")),
         ),
     ],
 )
 def test_generate_to_meta(
-    connectable,
-    schema_command,
+    dialect,
     schema,
     table_description,
     column_description,
     primary_key,
     float_type,
     bool_type,
-    postgres_connection,
 ):
-    if connectable == "postgres_engine":
-        connectable = postgres_connection[0]
 
-    create_schema(connectable, schema_command, schema)
+    connectable = create_sqlachemy_engine(dialect)
     create_tables(connectable, schema)
-
-    # To check the sqlalchemy data types:
-    # insp = inspect(connectable)
-    # logging.info(insp.get_table_comment(table_name,schema=schema))
-    # for i in insp.get_columns(table_name,schema):
-    #     logging.info(i)
-    #     logging.info(i["type"])
 
     sqlc = SQLAlchemyConverter(connectable)
     metadata = sqlc.generate_to_meta(table=table_name, schema=schema)
@@ -237,51 +225,7 @@ def test_generate_to_meta(
         bool_type=bool_type,
     )
 
-    # To check the mojap metadata types:
-    # for i in table_meta.to_dict()["columns"]:
-    #     logging.info(i)
-
-    # To check that the metadata objects can be converted:
-    # from mojap_metadata.converters.etl_manager_converter import EtlManagerConverter
-    # emc = EtlManagerConverter()
-    # my_lookup_dict = {"": "parquet"}
-    # etl_meta = emc.generate_from_meta(metadata=metadata,
-    #                                   file_format_mapper=my_lookup_dict.get)
-    # etl_meta.write_to_json(f"{connectable.dialect}.json")
-
     metadata_list = sqlc.generate_to_meta_list(schema=schema)
     assert len(metadata_list) == 2
     assert metadata_list[0].name == table_name
     assert metadata_list[1].name == "table2"
-
-
-@pytest.mark.parametrize(
-    "inputtype,expected",
-    [
-        (Integer(), "int32"),
-        (BIGINT(), "int64"),
-        (Float(precision=10, decimal_return_scale=2), "float64"),
-        (String(), "string"),
-        (String(length=4000), "string"),
-        (VARCHAR(length=255), "string"),
-        (NCHAR(length=10), "string"),
-        (CLOB(), "string"),
-        (Date(), "date64"),
-        (DateTime(), "datetime"),
-        (DATE(), "date64"),
-        (DATETIME(), "datetime"),
-        (TIMESTAMP(timezone=False), "timestamp(ms)"),
-        (Boolean(), "bool"),
-        (BLOB(), "binary"),
-        (VARBINARY(), "binary"),
-        (NUMERIC(precision=15, scale=2), "decimal128(15,2)"),
-        (DECIMAL(precision=8, scale=0), "decimal128(8,0)"),
-        (JSON(), "string"),
-        ("Unknown", "string"),
-    ],
-)
-def test_convert_to_mojap_type(inputtype: type, expected: str):
-    engine = create_engine("sqlite:///:memory:")
-    pc = SQLAlchemyConverter(engine)
-    actual = pc.convert_to_mojap_type(inputtype)
-    assert actual == expected
